requirements.dev.txt
djangorestframework>=3.7.0
humanize
flake8
django-nose
coverage
django-memoize
isort
pytest-django